refactor(frontend): shared/ui のファイル名を kebab-case へ統一する
shadcn/ui の生成物と同じファイル命名規約に揃えるため、既存の
共通 UI をリネームする。公開名(PascalCase)とバレル経由の参照は不変。

- AuthLayout.tsx → auth-layout.tsx
- ImageUpload.tsx → image-upload.tsx
- バレルの相対 import を新ファイル名へ更新
