+++ frontend/src/shared/ui/auth-layout.tsx
