+++ frontend/src/shared/ui/image-upload.tsx
